Revert "[iOS] Adding sentinel error histogram"
This reverts commit fdc27917d77732d07d228e2fa9ffd155c6e820a5.

Reason for revert: <INSERT REASONING HERE>

Original change's description:
> [iOS] Adding sentinel error histogram
> 
> Adding an histogram to log the error while creating the sentinel file.
> 
> TBR=jlebel@chromium.org
> 
> (cherry picked from commit 46c04be311f3f2885f0ad4e135cf985a2c019926)
> 
> Bug: 822806
> Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet;master.tryserver.chromium.mac:ios-simulator-full-configs
> Reviewed-on: https://chromium-review.googlesource.com/969261
> Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
> Cr-Original-Commit-Position: refs/heads/master@{#544776}
> Reviewed-on: https://chromium-review.googlesource.com/978204
> Cr-Commit-Position: refs/branch-heads/3359@{#393}
> Cr-Branched-From: 66afc5e5d10127546cc4b98b9117aff588b5e66b-refs/heads/master@{#540276}

TBR=gayane@chromium.org,pkl@chromium.org,jlebel@chromium.org,mark@chromium.org

No-Presubmit: true
No-Tree-Checks: true
No-Try: true
Bug: 822806
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet;master.tryserver.chromium.mac:ios-simulator-full-configs
Reviewed-on: https://chromium-review.googlesource.com/977972
Cr-Commit-Position: refs/branch-heads/3359@{#394}
Cr-Branched-From: 66afc5e5d10127546cc4b98b9117aff588b5e66b-refs/heads/master@{#540276}

+++ ios/chrome/browser/first_run/first_run.mm
@@ -61,16 +61,12 @@ bool FirstRun::RemoveSentinel() {
 }
 
 // static
-FirstRun::SentinelResult FirstRun::CreateSentinel(base::File::Error* error) {
+bool FirstRun::CreateSentinel() {
   base::FilePath first_run_sentinel;
-  if (!GetFirstRunSentinelFilePath(&first_run_sentinel))
-    return SENTINEL_RESULT_FAILED_TO_GET_PATH;
-  if (base::PathExists(first_run_sentinel))
-    return SENTINEL_RESULT_FILE_PATH_EXISTS;
-  bool success = base::WriteFile(first_run_sentinel, "", 0) != -1;
-  if (error)
-    *error = base::File::GetLastFileError();
-  return success ? SENTINEL_RESULT_SUCCESS : SENTINEL_RESULT_FILE_ERROR;
+  if (!GetFirstRunSentinelFilePath(&first_run_sentinel) ||
+      base::PathExists(first_run_sentinel))
+    return false;
+  return base::WriteFile(first_run_sentinel, "", 0) != -1;
 }
 
 // static
